# pilot: reduce lock contention in EDS