This PR reduces lock contention in EDS. Primarily discovered/discussed/tested in https://github.com/istio/istio/issues/44985. I was also able to reproduce the same

We see very little impact on the raw microbenchmarks:
```
name                          old time/op        new time/op        delta
EndpointGeneration/1/100-48          642µs ± 2%         653µs ± 2%  +1.78%  (p=0.002 n=9+10)
EndpointGeneration/10/10-48          300µs ± 1%         306µs ± 2%  +2.11%  (p=0.000 n=10+10)
EndpointGeneration/100/10-48        2.68ms ± 1%        2.69ms ± 1%    ~     (p=0.218 n=10+10)
EndpointGeneration/1000/1-48        2.66ms ± 1%        2.70ms ± 3%  +1.57%  (p=0.017 n=9+10)

name                          old kb/msg         new kb/msg         delta
EndpointGeneration/1/100-48           10.1 ± 0%          10.1 ± 0%    ~     (all equal)
EndpointGeneration/10/10-48           7.13 ± 0%          7.13 ± 0%    ~     (all equal)
EndpointGeneration/100/10-48          69.2 ± 0%          69.2 ± 0%    ~     (all equal)
EndpointGeneration/1000/1-48          69.6 ± 0%          69.6 ± 0%    ~     (all equal)

name                          old resources/msg  new resources/msg  delta
EndpointGeneration/1/100-48            100 ± 0%           100 ± 0%    ~     (all equal)
EndpointGeneration/10/10-48           10.0 ± 0%          10.0 ± 0%    ~     (all equal)
EndpointGeneration/100/10-48          10.0 ± 0%          10.0 ± 0%    ~     (all equal)
EndpointGeneration/1000/1-48          1.00 ± 0%          1.00 ± 0%    ~     (all equal)

name                          old alloc/op       new alloc/op       delta
EndpointGeneration/1/100-48          219kB ± 0%         220kB ± 0%  +0.37%  (p=0.000 n=10+10)
EndpointGeneration/10/10-48         79.9kB ± 0%        82.4kB ± 0%  +3.11%  (p=0.000 n=9+9)
EndpointGeneration/100/10-48         672kB ± 0%         693kB ± 0%  +3.03%  (p=0.000 n=10+9)
EndpointGeneration/1000/1-48         696kB ± 0%         722kB ± 0%  +3.63%  (p=0.000 n=10+10)

name                          old allocs/op      new allocs/op      delta
EndpointGeneration/1/100-48          4.41k ± 0%         4.51k ± 0%  +2.27%  (p=0.000 n=10+10)
EndpointGeneration/10/10-48          1.66k ± 0%         1.71k ± 0%  +3.02%  (p=0.000 n=10+10)
EndpointGeneration/100/10-48         12.7k ± 0%         12.8k ± 0%  +0.65%  (p=0.000 n=10+10)
EndpointGeneration/1000/1-48         12.1k ± 0%         12.2k ± 0%  +0.12%  (p=0.000 n=10+10)
```

The very slight regression is just due to computing a slice of the endpoints inside the lock so we can access them outside the lock, which we did not have before. 

The benchmark is an isolated test, so it doesn't exercise lock contention in any way, hence why it looks strictly worse there. In my real world testing, this completely removed lock contention