@istio-testing 2023-05-31
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
unit-tests-arm64_istio | 844ae885925c445bd066f421d60730c444e2018a | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45224/unit-tests-arm64_istio/1663977635882995712) | true | `/test unit-tests-arm64`
lint_istio | 844ae885925c445bd066f421d60730c444e2018a | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45224/lint_istio/1663977646700105728) | true | `/test lint`
unit-tests_istio | 844ae885925c445bd066f421d60730c444e2018a | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/45224/unit-tests_istio/1663977635820081152) | true | `/test unit-tests`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->