@ra-grover 2023-05-31
Thanks for opening this @howardjohn , and glad that you were able to reproduce the contention. Is there any chance that this can be backported to v1.17.1 ?