@howardjohn 2023-06-06
it's very hard to have a benchmark for lock contention. I don't know any
reliable way to test it as it only occurs when CPU is throttling

On Mon, Jun 5, 2023, 7:37 PM Zhonghu Xu ***@***.***> wrote:

> ***@***.**** commented on this pull request.
>
> BTW: can we add a benchmark for it
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/istio/istio/pull/45224#pullrequestreview-1464117179>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAEYGXKHKY4XZG3LUPG5MNTXJ2JW3ANCNFSM6AAAAAAYV3P6EA>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>